And everything is going to the beat
