--- src/main/java/org/spongepowered/api/item/recipe/CraftingRecipe.java
@@ -1,5 +1,0 @@
-package org.spongepowered.api.item.recipe;
-
-public interface CraftingRecipe extends Recipe {
-
-}
